Got the Status to not be null anymore. Now trying to make it so the product is no longer null...

+++ src/main/java/com/codewithmosh/dtos/OrderProductDto.java
@@ -1,11 +1,14 @@
 package com.codewithmosh.dtos;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 
 public class OrderProductDto {
-    private long id;
+    private Long id;
     private String name;
-    private double price;
+    private BigDecimal price;
 }
